SWS-258 Refactor Service Details Info
