+++ src/pages/ServiceDetails/ServiceInfo/index.tsx
@@ -1,0 +1,15 @@
+import { default as ServiceInfoPods } from './ServiceInfoPods';
+import { default as ServiceInfoDescription } from './ServiceInfoDescription';
+import { default as ServiceInfoRoutes } from './ServiceInfoRoutes';
+import { default as ServiceInfoRules } from './ServiceInfoRules';
+import { default as ServiceInfoBadge } from './ServiceInfoBadge';
+import { default as ServiceInfoCard } from './ServiceInfoCard';
+
+export {
+  ServiceInfoPods,
+  ServiceInfoDescription,
+  ServiceInfoRoutes,
+  ServiceInfoRules,
+  ServiceInfoBadge,
+  ServiceInfoCard
+};
